+++ src/device-types/HASelect.cpp
@@ -85,6 +85,11 @@ bool HASelect::setState(const int8_t state, const bool force)
     return false;
 }
 
+const char* HASelect::getCurrentOption() const
+{
+    return _options->getItem(getCurrentState());
+}
+
 void HASelect::buildSerializer()
 {
     if (_serializer || !uniqueId() || !_options) {
@@ -164,8 +169,12 @@ void HASelect::onMqttMessage(
 
 bool HASelect::publishState(const int8_t state)
 {
-    if (state == -1 || !_options || state >= _options->getItemsNb()) {
+    if (!_options || state >= _options->getItemsNb()) {
         return false;
+    }
+
+    if (state == -1) {
+        return publishOnDataTopic(AHATOFSTR(HAStateTopic), AHATOFSTR(HAStateNone), true);
     }
 
     const char* item = _options->getItems()[state];